chore: cleanup and simplify worker creation
