+++ pooler/utils/callback_helpers.py
@@ -4,8 +4,16 @@ import multiprocessing
 import resource
 import signal
 import time
+from abc import ABC
+from abc import ABCMeta
+from abc import abstractmethod
+from abc import abstractproperty
 from functools import partial
+from functools import wraps
+from typing import Callable
 from typing import Dict
+from typing import List
+from typing import Optional
 from typing import Tuple
 from typing import Union
 from urllib.parse import urljoin
@@ -16,6 +24,10 @@ from aio_pika import IncomingMessage
 from aio_pika.pool import Pool
 from eth_utils import keccak
 from httpx import AsyncClient
+from httpx import AsyncHTTPTransport
+from httpx import Limits
+from httpx import Timeout
+from pydantic import ValidationError
 from redis import asyncio as aioredis
 from setproctitle import setproctitle
 from tenacity import AsyncRetrying
@@ -25,13 +37,18 @@ from tenacity import wait_random_exponential
 from pooler.settings.config import settings
 from pooler.utils.default_logger import logger
 from pooler.utils.models.data_models import PayloadCommitAPIRequest
+from pooler.utils.models.data_models import SnapshotterIssue
+from pooler.utils.models.data_models import SnapshotterIssueSeverity
 from pooler.utils.models.data_models import SourceChainDetails
 from pooler.utils.models.message_models import PowerloomCallbackProcessMessage
 from pooler.utils.models.message_models import SnapshotBase
+from pooler.utils.redis.rate_limiter import load_rate_limiter_scripts
 from pooler.utils.redis.redis_conn import RedisPoolCache
 from pooler.utils.redis.redis_keys import (
     cb_broadcast_processing_logs_zset,
 )
+from pooler.utils.redis.redis_keys import discarded_query_epochs_redis_q
+from pooler.utils.redis.redis_keys import failed_query_epochs_redis_q
 
 # setup logger
 helper_logger = logger.bind(module='PowerLoom|Callback|Helpers')
@@ -50,6 +67,43 @@ async def get_rabbitmq_connection():
 async def get_rabbitmq_channel(connection_pool) -> aio_pika.Channel:
     async with connection_pool.acquire() as connection:
         return await connection.channel()
+
+
+def notify_on_task_failure(fn):
+    @wraps(fn)
+    async def wrapper(self, *args, **kwargs):
+        try:
+            await fn(self, *args, **kwargs)
+
+        except Exception as e:
+            # Logging the error trace
+            logger.opt(exception=True).error(f'Error: {e}')
+
+            # Sending the error details to the issue reporting service
+            try:
+                if 'msg_obj' in kwargs:
+                    msg_obj = kwargs['msg_obj']
+                    if isinstance(msg_obj, PowerloomCallbackProcessMessage):
+                        contract = msg_obj.contract
+                        project_id = f'uniswap_pairContract_*_{contract}_{settings.namespace}'
+
+                await self._client.post(
+                    url=settings.issue_report_url,
+                    json=SnapshotterIssue(
+                        instanceID=settings.instance_id,
+                        severity=SnapshotterIssueSeverity.medium,
+                        issueType='MISSED_SNAPSHOT',
+                        projectID=project_id if project_id else '*',
+                        timeOfReporting=int(time.time()),
+                        extra={'issueDetails': f'Error : {e}'},
+                        serviceName='Pooler|PairTotalReservesProcessor',
+                    ).dict(),
+                )
+            except Exception as err:
+                # Logging the error trace if service is not able to report issue
+                logger.opt(exception=True).error(f'Error: {err}')
+
+    return wrapper
 
 
 class AuditProtocolCommandsHelper:
@@ -100,7 +154,9 @@ class AuditProtocolCommandsHelper:
                     )
 
 
-class CallbackAsyncWorker(multiprocessing.Process):
+class CallbackAsyncWorker(multiprocessing.Process, ABC):
+    __metaclass__ = ABCMeta
+
     def __init__(self, name, rmq_q, rmq_routing, **kwargs):
         self._core_rmq_consumer: asyncio.Task
         self._q = rmq_q
@@ -112,7 +168,41 @@ class CallbackAsyncWorker(multiprocessing.Process):
         super(CallbackAsyncWorker, self).__init__(name=name, **kwargs)
         self._logger = logger.bind(module=self.name)
 
+        self._rate_limiting_lua_scripts = dict()
+        self._async_transport = AsyncHTTPTransport(
+            limits=Limits(
+                max_connections=100,
+                max_keepalive_connections=50,
+                keepalive_expiry=None,
+            ),
+        )
+        self._client = AsyncClient(
+            timeout=Timeout(timeout=5.0),
+            follow_redirects=False,
+            transport=self._async_transport,
+        )
         self._shutdown_signal_received_count = 0
+
+    @abstractproperty
+    def _snapshot_name(self):
+        pass
+
+    @abstractproperty
+    def _stream(self):
+        pass
+
+    @abstractproperty
+    def _transformation_lambdas(self):
+        pass
+
+    @abstractmethod
+    async def _compute(
+        self,
+        min_chain_height: int,
+        max_chain_height: int,
+        data_source_contract_address: str,
+    ):
+        pass
 
     async def _shutdown_handler(self, sig, loop: asyncio.AbstractEventLoop):
         self._shutdown_signal_received_count += 1
@@ -189,6 +279,66 @@ class CallbackAsyncWorker(multiprocessing.Process):
             await asyncio.gather(*tasks, return_exceptions=True)
             loop.stop()
             self._logger.info('Shutdown complete.')
+
+    @notify_on_task_failure
+    async def _processor_task(self, msg_obj: PowerloomCallbackProcessMessage):
+        """Function used to process the received message object."""
+        self._logger.debug(
+            'Processing total pair reserves callback: {}', msg_obj,
+        )
+
+        self_unique_id = str(uuid4())
+        cur_task: asyncio.Task = asyncio.current_task(
+            asyncio.get_running_loop(),
+        )
+        cur_task.set_name(
+            f'aio_pika.consumer|PairTotalReservesProcessor|{msg_obj.contract}',
+        )
+        self._running_callback_tasks[self_unique_id] = cur_task
+
+        await self.init_redis_pool()
+        if not self._rate_limiting_lua_scripts:
+            self._rate_limiting_lua_scripts = await load_rate_limiter_scripts(
+                self._redis_conn,
+            )
+        self._logger.debug(
+            'Got epoch to process for calculating total reserves for pair: {}',
+            msg_obj,
+        )
+
+        # check for enqueued failed query epochs
+        epochs = await self._prepare_epochs(
+            failed_query_epochs_key=failed_query_epochs_redis_q.format(
+                self._stream, msg_obj.contract,
+            ),
+            discarded_query_epochs_key=discarded_query_epochs_redis_q.format(
+                self._stream, msg_obj.contract,
+            ),
+            current_epoch=msg_obj,
+            snapshot_name=self._snapshot_name,
+            failed_query_epochs_l=[],
+            stream=self._stream,
+        )
+
+        epoch_snapshot_map = await self._map_processed_epochs_to_adapters(
+            epochs=epochs,
+            cb_fn_async=self._compute,
+            enqueue_on_failure=True,
+            data_source_contract_address=msg_obj.contract,
+            failed_query_epochs_key=failed_query_epochs_redis_q.format(
+                self._stream, msg_obj.contract,
+            ),
+            transformation_lambdas=self._transformation_lambdas,
+        )
+
+        await self._send_audit_payload_commit_service(
+            audit_stream=self._stream,
+            original_epoch=msg_obj,
+            snapshot_name=self._snapshot_name,
+            epoch_snapshot_map=epoch_snapshot_map,
+        )
+
+        del self._running_callback_tasks[self_unique_id]
 
     async def _send_audit_payload_commit_service(
         self,
@@ -339,6 +489,202 @@ class CallbackAsyncWorker(multiprocessing.Process):
                         mapping={json.dumps(update_log): int(time.time())},
                     )
 
+    async def _prepare_epochs(
+        self,
+        failed_query_epochs_key: str,
+        stream: str,
+        discarded_query_epochs_key: str,
+        current_epoch: PowerloomCallbackProcessMessage,
+        snapshot_name: str,
+        failed_query_epochs_l: Optional[List],
+    ) -> List[PowerloomCallbackProcessMessage]:
+        queued_epochs = list()
+        # checks for any previously queued epochs, returns a list of such epochs in increasing order of blockheights
+        if settings.env != 'test':
+            project_id = (
+                f'{stream}_{current_epoch.contract}_{settings.namespace}'
+            )
+            fall_behind_reset_threshold = (
+                settings.rpc.skip_epoch_threshold_blocks
+            )
+            failed_query_epochs = await self._redis_conn.lpop(
+                failed_query_epochs_key,
+            )
+            while failed_query_epochs:
+                epoch_broadcast: PowerloomCallbackProcessMessage = (
+                    PowerloomCallbackProcessMessage.parse_raw(
+                        failed_query_epochs.decode('utf-8'),
+                    )
+                )
+                if (
+                    current_epoch.begin - epoch_broadcast.end >
+                    fall_behind_reset_threshold
+                ):
+                    # send alert
+                    await self._client.post(
+                        url=settings.issue_report_url,
+                        json=SnapshotterIssue(
+                            instanceID=settings.instance_id,
+                            severity=SnapshotterIssueSeverity.medium,
+                            issueType='SKIP_QUEUED_EPOCH',
+                            projectID=project_id,
+                            epochs=[epoch_broadcast.end],
+                            timeOfReporting=int(time.time()),
+                            serviceName=f'Pooler|CallbackProcessor|{stream}',
+                        ).dict(),
+                    )
+                    await self._redis_conn.rpush(
+                        discarded_query_epochs_key,
+                        epoch_broadcast.json(),
+                    )
+                    self._logger.warning(
+                        (
+                            'Project {} | QUEUED Epoch {} processing has fallen'
+                            ' behind by more than {} blocks, alert sent to DAG'
+                            ' Verifier | Discarding queued epoch'
+                        ),
+                        project_id,
+                        epoch_broadcast,
+                        fall_behind_reset_threshold,
+                    )
+                else:
+                    self._logger.info(
+                        (
+                            'Found queued epoch against which snapshot'
+                            " construction for pair contract's {} failed"
+                            ' earlier: {}'
+                        ),
+                        snapshot_name,
+                        epoch_broadcast,
+                    )
+                    queued_epochs.append(epoch_broadcast)
+                failed_query_epochs = await self._redis_conn.lpop(
+                    failed_query_epochs_key,
+                )
+        else:
+            queued_epochs = (
+                failed_query_epochs_l if failed_query_epochs_l else list()
+            )
+        queued_epochs.append(current_epoch)
+        # check for continuity in epochs before ordering them
+        self._logger.info(
+            (
+                'Attempting to check for continuity in queued epochs to'
+                " generate snapshots against pair contract's {} including"
+                ' current epoch: {}'
+            ),
+            snapshot_name,
+            queued_epochs,
+        )
+        continuity = True
+        for idx, each_epoch in enumerate(queued_epochs):
+            if idx == 0:
+                continue
+            if each_epoch.begin != queued_epochs[idx - 1].end + 1:
+                continuity = False
+                break
+        if not continuity:
+            # mark others as discarded
+            discarded_epochs = queued_epochs[:-1]
+            # pop off current epoch added to end of this list
+            queued_epochs = [queued_epochs[-1]]
+            self._logger.info(
+                (
+                    'Recording epochs as discarded during snapshot construction'
+                    ' stage for {}: {}'
+                ),
+                snapshot_name,
+                queued_epochs,
+            )
+            for x in discarded_epochs:
+                await self._redis_conn.rpush(
+                    discarded_query_epochs_key,
+                    x.json(),
+                )
+        return queued_epochs
+
+    async def _map_processed_epochs_to_adapters(
+        self,
+        epochs: List[PowerloomCallbackProcessMessage],
+        cb_fn_async,
+        enqueue_on_failure,
+        data_source_contract_address,
+        failed_query_epochs_key,
+        transformation_lambdas: List[Callable],
+    ):
+        tasks_map = dict()
+        for each_epoch in epochs:
+            tasks_map[
+                (
+                    each_epoch.begin,
+                    each_epoch.end,
+                    each_epoch.broadcast_id,
+                )
+            ] = cb_fn_async(
+                min_chain_height=each_epoch.begin,
+                max_chain_height=each_epoch.end,
+                data_source_contract_address=data_source_contract_address,
+            )
+        results = await asyncio.gather(
+            *tasks_map.values(), return_exceptions=True,
+        )
+        results_map = dict()
+        for idx, each_result in enumerate(results):
+            epoch_against_result = list(tasks_map.keys())[idx]
+            if (
+                isinstance(each_result, Exception) and
+                enqueue_on_failure and
+                settings.env != 'test'
+            ):
+                queue_msg_obj = PowerloomCallbackProcessMessage(
+                    begin=epoch_against_result[0],
+                    end=epoch_against_result[1],
+                    broadcast_id=epoch_against_result[2],
+                    contract=data_source_contract_address,
+                )
+                await self._redis_conn.rpush(
+                    failed_query_epochs_key,
+                    queue_msg_obj.json(),
+                )
+                self._logger.debug(
+                    (
+                        'Enqueued epoch broadcast ID {} because reserve query'
+                        ' failed on {} - {} | Exception: {}'
+                    ),
+                    queue_msg_obj.broadcast_id,
+                    epoch_against_result[0],
+                    epoch_against_result[1],
+                    each_result,
+                )
+                results_map[
+                    (epoch_against_result[0], epoch_against_result[1])
+                ] = None
+            else:
+                if not isinstance(each_result, Exception):
+                    for transformation in transformation_lambdas:
+                        each_result = transformation(
+                            each_result,
+                            data_source_contract_address,
+                            epoch_against_result[0],
+                            epoch_against_result[1],
+                        )
+                    results_map[
+                        (
+                            epoch_against_result[0],
+                            epoch_against_result[1],
+                        )
+                    ] = each_result
+            return results_map
+
+    async def _update_broadcast_processing_status(
+        self, broadcast_id, update_state,
+    ):
+        await self._redis_conn.hset(
+            cb_broadcast_processing_logs_zset.format(self.name),
+            broadcast_id,
+            json.dumps(update_state),
+        )
+
     async def _rabbitmq_consumer(self, loop):
         self._rmq_connection_pool = Pool(
             get_rabbitmq_connection, max_size=5, loop=loop,
@@ -364,6 +710,31 @@ class CallbackAsyncWorker(multiprocessing.Process):
 
     async def _on_rabbitmq_message(self, message: IncomingMessage):
         await message.ack()
+
+        try:
+            msg_obj: PowerloomCallbackProcessMessage = (
+                PowerloomCallbackProcessMessage.parse_raw(message.body)
+            )
+        except ValidationError as e:
+            self._logger.opt(exception=True).error(
+                (
+                    'Bad message structure of callback in processor for total'
+                    ' pair reserves: {}'
+                ),
+                e,
+            )
+            return
+        except Exception as e:
+            self._logger.opt(exception=True).error(
+                (
+                    'Unexpected message structure of callback in processor for'
+                    ' total pair reserves: {}'
+                ),
+                e,
+            )
+            return
+
+        asyncio.ensure_future(self._processor_task(msg_obj=msg_obj))
 
     async def init_redis_pool(self):
         if not self._aioredis_pool:
